+++ cartridge/webui/api-config.lua
@@ -4,6 +4,11 @@ local log = require('log')
 local json = require('json').new()
 local yaml = require('yaml').new()
 local errors = require('errors')
+local checks = require('checks')
+
+local utils = require('cartridge.utils')
+local gql_types = require('cartridge.graphql.types')
+local module_name = 'cartridge.webui.api-config'
 
 json.cfg({
     encode_use_tostring = true,
@@ -12,6 +17,9 @@ yaml.cfg({
     encode_use_tostring = true,
 })
 
+local DecodeYamlError = errors.new_class('DecodeYamlError')
+local DownloadConfigError = errors.new_class('Config download failed')
+local UploadConfigError = errors.new_class('Config upload failed')
 
 local auth = require('cartridge.auth')
 local twophase = require('cartridge.twophase')
@@ -30,6 +38,24 @@ local system_sections = {
     ['vshard_groups.yml'] = true,
 }
 
+local gql_type_section = gql_types.object({
+    name = 'ConfigSection',
+    description = 'A section of clusterwide configuration',
+    fields = {
+        filename = gql_types.string.nonNull,
+        content = gql_types.string.nonNull
+    }
+})
+
+local gql_type_section_input = gql_types.inputObject({
+    name = 'ConfigSectionInput',
+    description = 'A section of clusterwide configuration',
+    fields = {
+        filename = gql_types.string.nonNull,
+        content = gql_types.string
+    }
+})
+
 local function http_finalize_error(http_code, err)
     log.error('%s', err)
 
@@ -42,16 +68,15 @@ local function http_finalize_error(http_code, err)
     })
 end
 
-local e_download_config = errors.new_class('Config download failed')
 local function download_config_handler(req)
     if not auth.authorize_request(req) then
-        local err = e_download_config:new('Unauthorized')
+        local err = DownloadConfigError:new('Unauthorized')
         return http_finalize_error(401, err)
     end
 
     local clusterwide_config = confapplier.get_active_config()
     if clusterwide_config == nil then
-        local err = e_download_config:new("Cluster isn't bootsrapped yet")
+        local err = DownloadConfigError:new("Cluster isn't bootstrapped yet")
         return http_finalize_error(409, err)
     end
 
@@ -81,16 +106,14 @@ local function download_config_handler(req)
     })
 end
 
-local e_upload_config = errors.new_class('Config upload failed')
-local e_decode_yaml = errors.new_class('Decoding YAML failed')
 local function upload_config_handler(req)
     if not auth.authorize_request(req) then
-        local err = e_upload_config:new('Unauthorized')
+        local err = UploadConfigError:new('Unauthorized')
         return http_finalize_error(401, err)
     end
 
     if confapplier.get_readonly() == nil then
-        local err = e_upload_config:new("Cluster isn't bootsrapped yet")
+        local err = UploadConfigError:new("Cluster isn't bootstrapped yet")
         return http_finalize_error(409, err)
     end
 
@@ -118,17 +141,17 @@ local function upload_config_handler(req)
         req_body = form_body
     end
 
-    local conf_new, err = nil
+    local conf_new, err
     if req_body == nil then
-        err = e_upload_config:new('Request body must not be empty')
+        conf_new, err = nil, UploadConfigError:new('Request body must not be empty')
     else
-        conf_new, err = e_decode_yaml:pcall(yaml.decode, req_body)
+        conf_new, err = DecodeYamlError:pcall(yaml.decode, req_body)
     end
 
     if err ~= nil then
         return http_finalize_error(400, err)
     elseif type(conf_new) ~= 'table' then
-        err = e_upload_config:new('Config must be a table')
+        err = UploadConfigError:new('Config must be a table')
         return http_finalize_error(400, err)
     end
 
@@ -145,12 +168,12 @@ local function upload_config_handler(req)
 
     for k, v in pairs(conf_new) do
         if system_sections[k] then
-            local err = e_upload_config:new(
+            local err = UploadConfigError:new(
                 "uploading system section %q is forbidden", k
             )
             return http_finalize_error(400, err)
         elseif type(conf_new[k .. '.yml']) ~= 'nil' then
-            local err = e_upload_config:new(
+            local err = UploadConfigError:new(
                 "ambiguous sections %q and %q", k, k .. '.yml'
             )
             return http_finalize_error(400, err)
@@ -169,7 +192,85 @@ local function upload_config_handler(req)
     return auth.render_response({status = 200})
 end
 
-local function init(httpd)
+local function get_sections(_, args)
+    checks('?', {sections = '?table'})
+    local clusterwide_config = confapplier.get_active_config()
+    if clusterwide_config == nil then
+        local err = DownloadConfigError:new("Cluster isn't bootstrapped yet")
+        return nil, err
+    end
+
+    local ret = {}
+    for section, content in pairs(clusterwide_config:get_plaintext()) do
+        if (args.sections == nil or utils.table_find(args.sections, section))
+        and not system_sections[section] then
+            table.insert(ret, {
+                filename = section,
+                content = content,
+            })
+        end
+    end
+
+    return ret
+end
+
+local function set_sections(_, args)
+    checks('?', {sections = '?table'})
+    if confapplier.get_readonly() == nil then
+        local err = UploadConfigError:new("Cluster isn't bootstrapped yet")
+        return nil, err
+    end
+
+    if args.sections == nil then
+        args.sections = {}
+    end
+
+    local patch = {}
+    local query_sections = {}
+
+    for _, input in ipairs(args.sections) do
+        if system_sections[input.filename] then
+            return nil, UploadConfigError:new(
+                "uploading system section %q is forbidden",
+                input.filename
+            )
+        end
+
+        patch[input.filename] = input.content or box.NULL
+        table.insert(query_sections, input.filename)
+    end
+
+    local ok, err = twophase.patch_clusterwide(patch)
+    if not ok then
+        return nil, err
+    end
+
+    return get_sections(nil, {sections = query_sections})
+end
+
+
+local function init(graphql, httpd)
+    graphql.add_callback({
+        prefix = 'cluster',
+        name = 'config',
+        doc = 'Get cluster config sections',
+        args = {
+            sections = gql_types.list(gql_types.string.nonNull)
+        },
+        kind = gql_types.list(gql_type_section).nonNull,
+        callback = module_name .. '.get_sections',
+    })
+
+    graphql.add_mutation({
+        prefix = 'cluster',
+        name = 'config',
+        doc = 'Applies updated config on cluster',
+        args = {
+            sections = gql_types.list(gql_type_section_input)
+        },
+        kind = gql_types.list(gql_type_section).nonNull,
+        callback = module_name .. '.set_sections',
+    })
 
     httpd:route({
         path = '/admin/config',
@@ -185,4 +286,6 @@ end
 
 return {
     init = init,
+    get_sections = get_sections,
+    set_sections = set_sections,
 }